+++ src/globals.d.ts
@@ -2,3 +2,5 @@ declare const __APP_VERSION__: string;
 declare const __REMOTE_JSON_URL__: string;
 declare const __REMOTE_APP_RELEASE_URL__: string;
 declare const __GITHUB_RELEASES_API_URL__: string;
+declare const __DOWNLOAD_GATEWAY_URL__: string;
+declare const __DOWNLOAD_GATEWAY_STRICT__: boolean;